automated test on enforcing https
created automated test for verification of the correct creation of the and setup of https.env file when the user opts to enforce-https

This code creates a test case that runs a single helper function
within the typer framework to validate the outputs and side-effects of
running it.  The function under test is a helper function in the
actual deployment script, but we want to test it on its own, so this
approach uses a lambda to bind the input variables of the
function so it can be invoked by typer as if it were the main
callback (which takes no inputs except the actual command line
parameters).

remove extra parentheses

remove extra space

Update init-odkx-sync-endpoint.py