add ability to support Internationalisation
This commit uses python's locale module to get the default language set in your system and the default gettext module to load the messages basing on the provided locale.  In case there is no default language set, English is taken as default and the user is notified about this.

solved french not working

solved mypy issue

remove pytest cache